Draw app finished
